requirements.txt
# ===== Sanal Öğretmen Asistanı - Bağımlılıklar =====

# --- LLM & Orkestrasyon ---
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0

# --- RAG Pipeline ---
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
pdfplumber>=0.11.0

# --- Web Arayüzü ---
streamlit>=1.38.0

# --- Derin Öğrenme (DKT & DRL) ---
torch>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0

# --- Veri & Ortam ---
python-dotenv>=1.0.0
pandas>=2.2.0

# --- Geliştirme Araçları ---
pytest>=8.0.0
